backend_ML/requirements.txt
# --- Core FastAPI Dependencies ---
fastapi==0.115.2
uvicorn==0.31.1
python-multipart==0.0.9
aiofiles==24.1.0
jinja2==3.1.4

# --- File Handling & Parsing ---
python-docx==1.1.2
pymupdf==1.24.9

# --- Environment & Requests ---
python-dotenv==1.0.1
requests==2.32.3

# --- Google Gemini AI ---
google-generativeai==0.8.3

# --- Optional (Logging & Async Utilities) ---
loguru==0.7.2
youtube-transcript-api==0.6.2